Massive reduction and increase in robustness of server update structure
