+++ server/container.tokens.js
@@ -6,6 +6,7 @@ const DITokens = {
   serverHost: "serverHost",
   loggerFactory: "loggerFactory",
   httpClient: "httpClient",
+  simpleGitService: "simpleGitService",
   multerService: "multerService",
   taskManagerService: "taskManagerService",
   toadScheduler: "toadScheduler",
@@ -13,6 +14,7 @@ const DITokens = {
   printerService: "printerService",
   printerGroupService: "printerGroupService",
   serverSettingsService: "serverSettingsService",
+  serverReleaseService: "serverReleaseService",
   serverUpdateService: "serverUpdateService",
   clientSettingsService: "clientSettingsService",
   githubApiService: "githubApiService",
@@ -59,7 +61,6 @@ const DITokens = {
   printerSseTask: "printerSseTask",
   printerSseHandler: "printerSseHandler",
   printerTestTask: "printerTestTask",
-  systemCommandsService: "systemCommandsService",
   printerWebsocketTask: "printerWebsocketTask",
   printerWebsocketPingTask: "printerWebsocketPingTask",
   printerFileCleanTask: "printerFileCleanTask"
