--- server/controllers/validation/server-update.validation.js
@@ -1,8 +1,0 @@
-const updateServerRules = {
-    forcePull: "required|boolean",
-    installPackageUpdates: "required|boolean"
-};
-
-module.exports = {
-    updateServerRules
-};
