+++ server/package.json
@@ -2,7 +2,7 @@
   "name": "fdm-monster",
   "author": "David Zwart",
   "license": "AGPLv3",
-  "version": "1.1.0",
+  "version": "1.1.1",
   "description": "FDM Monster is a bulk OctoPrint manager to set up, configure and monitor 3D printers. Our aim is to provide extremely optimized websocket performance and reliability.",
   "main": "index.js",
   "scripts": {
@@ -56,7 +56,6 @@
     "influx": "^5.9.3",
     "is-docker": "^2.2.1",
     "lodash": "^4.17.21",
-    "lookpath": "^1.2.2",
     "luxon": "^2.3.1",
     "migrate-mongo": "^8.2.3",
     "mongoose": "^6.2.3",
@@ -70,7 +69,7 @@
     "rxjs-websockets": "^9.0.0",
     "semver": "^7.3.5",
     "simple-git": "^3.2.6",
-    "systeminformation": "^5.11.4",
+    "systeminformation": "^5.11.5",
     "toad-scheduler": "^1.6.0",
     "winston": "^3.6.0",
     "ws": "^8.5.0",
